dbcreate.c: (number_dnames_iterator) Number of dnames is stored in uint32_t, not in size_t (caused Bus Error on Solaris SPARC).
git-svn-id: file:///svn/nsd/trunk@950 a26ef69c-88ff-0310-839f-98b793d9c207

+++ dbcreate.c
@@ -209,7 +209,7 @@ write_rrset(struct namedb *db, domain_type *domain, rrset_type *rrset)
 static void
 number_dnames_iterator(domain_type *node, void *user_data)
 {
-	size_t *current_number = user_data;
+	uint32_t *current_number = user_data;
 
 	node->number = *current_number;
 	++*current_number;
